# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API 文档",
        "description": "Laravel RESTful API 文档",
        "contact": {
            "name": "API 支持团队",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API 服务器"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "客户管理"
                ],
                "summary": "获取所有客户",
                "description": "显示客户列表",
                "operationId": "391d8aab1cfed2de68caed90bfee32d7",
                "responses": {
                    "200": {
                        "description": "成功获取客户列表"
                    }
                }
            },
            "post": {
                "tags": [
                    "客户管理"
                ],
                "summary": "创建新客户",
                "description": "存储新客户",
                "operationId": "63cebef8fe6d9fe864316ed7b07529b4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "客户创建成功"
                    },
                    "422": {
                        "description": "验证错误"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "客户管理"
                ],
                "summary": "获取指定客户",
                "description": "显示指定客户",
                "operationId": "6a3a328b3e7caebbd82fb572706a731a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取客户信息"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            },
            "put": {
                "tags": [
                    "客户管理"
                ],
                "summary": "更新客户信息",
                "description": "更新指定客户",
                "operationId": "a14795ba8c398bf555801cfbc767a8ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "客户更新成功"
                    },
                    "422": {
                        "description": "验证错误"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            },
            "delete": {
                "tags": [
                    "客户管理"
                ],
                "summary": "删除客户",
                "description": "删除指定客户",
                "operationId": "5af7820dfd71770b5c79879a0f12bce7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "客户删除成功"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "显示所有客户列表",
                "description": "显示所有客户列表",
                "operationId": "30ce4124bc07a0d013b69974ff225a5c",
                "responses": {
                    "200": {
                        "description": "显示客户列表页面"
                    }
                }
            },
            "post": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "创建新客户",
                "description": "创建客户",
                "operationId": "8350a1a36d2b5522ea4158253eff28a1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "创建成功重定向"
                    }
                }
            }
        },
...
        "/customers/{id}": {
            "get": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "显示指定客户详情",
                "description": "显示指定客户",
                "operationId": "bb85afd644c1b965544c2f0e51f408b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "显示客户详情页面"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            },
            "put": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "更新客户信息",
                "description": "更新指定客户",
                "operationId": "7f856cee51366cc0a4c5e1aa6a02547a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "更新成功重定向"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            },
            "delete": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "删除客户",
                "description": "删除指定客户",
                "operationId": "4cee47d6b826da6ed7d852b5aff46d19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "删除成功重定向"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            }
        },
        "/customers/{id}/edit": {
            "get": {
                "tags": [
                    "客户管理-Web"
                ],
                "summary": "显示编辑客户表单",
                "description": "显示编辑客户表单",
                "operationId": "3f43a8cd7ab65554d1e511c7f6ee0bc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "显示编辑客户表单"
                    },
                    "404": {
                        "description": "客户不存在"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}